+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/RxDocumentClientImpl.java
@@ -1580,7 +1580,7 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
             }
 
             Instant serializationStartTime = Instant.now();
-            partitionKeyInternal =  extractPartitionKeyValueFromDocument(internalObjectNode, partitionKeyDefinition);
+            partitionKeyInternal =  PartitionKeyHelper.extractPartitionKeyValueFromDocument(internalObjectNode, partitionKeyDefinition);
             Instant serializationEndTime = Instant.now();
             SerializationDiagnosticsContext.SerializationDiagnostics serializationDiagnostics = new SerializationDiagnosticsContext.SerializationDiagnostics(
                 serializationStartTime,
@@ -1598,44 +1598,6 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
 
         request.setPartitionKeyInternal(partitionKeyInternal);
         request.getHeaders().put(HttpConstants.HttpHeaders.PARTITION_KEY, Utils.escapeNonAscii(partitionKeyInternal.toJson()));
-    }
-
-    public static PartitionKeyInternal extractPartitionKeyValueFromDocument(
-            JsonSerializable document,
-            PartitionKeyDefinition partitionKeyDefinition) {
-        if (partitionKeyDefinition != null) {
-            switch (partitionKeyDefinition.getKind()) {
-                case HASH:
-                    String path = partitionKeyDefinition.getPaths().iterator().next();
-                    List<String> parts = PathParser.getPathParts(path);
-                    if (parts.size() >= 1) {
-                        Object value = ModelBridgeInternal.getObjectByPathFromJsonSerializable(document, parts);
-                        if (value == null || value.getClass() == ObjectNode.class) {
-                            value = ModelBridgeInternal.getNonePartitionKey(partitionKeyDefinition);
-                        }
-
-                        if (value instanceof PartitionKeyInternal) {
-                            return (PartitionKeyInternal) value;
-                        } else {
-                            return PartitionKeyInternal.fromObjectArray(Collections.singletonList(value), false);
-                        }
-                    }
-                    break;
-                case MULTI_HASH:
-                    Object[] partitionKeyValues = new Object[partitionKeyDefinition.getPaths().size()];
-                    for(int pathIter = 0 ; pathIter < partitionKeyDefinition.getPaths().size(); pathIter++){
-                        String partitionPath = partitionKeyDefinition.getPaths().get(pathIter);
-                        List<String> partitionPathParts = PathParser.getPathParts(partitionPath);
-                        partitionKeyValues[pathIter] = ModelBridgeInternal.getObjectByPathFromJsonSerializable(document, partitionPathParts);
-                    }
-                    return PartitionKeyInternal.fromObjectArray(partitionKeyValues, false);
-
-                default:
-                    throw new IllegalArgumentException("Unrecognized Partition kind: " + partitionKeyDefinition.getKind());
-                }
-        }
-
-        return null;
     }
 
     private Mono<RxDocumentServiceRequest> getCreateDocumentRequest(DocumentClientRetryPolicy requestRetryPolicy,
@@ -1721,11 +1683,6 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
             requestHeaders,
             options,
             content);
-
-        if (serverBatchRequest instanceof PartitionKeyRangeServerBatchRequest) {
-            PartitionKeyRangeServerBatchRequest partitionKeyRangeServerBatchRequest = (PartitionKeyRangeServerBatchRequest) serverBatchRequest;
-            request.setPartitionBasedGoneNotifier(partitionKeyRangeServerBatchRequest.getPartitionBasedGoneNotifier());
-        }
 
         if (options != null) {
             request.requestContext.setExcludeRegions(options.getExcludeRegions());
@@ -4588,7 +4545,11 @@ public class RxDocumentClientImpl implements AsyncDocumentClient, IAuthorization
                     this.connectionPolicy.getConnectionMode(),
                     this.partitionKeyRangeCache);
 
-            this.storeModel.enableThroughputControl(throughputControlStore);
+            if (ConnectionMode.DIRECT == this.connectionPolicy.getConnectionMode()) {
+                this.storeModel.enableThroughputControl(throughputControlStore);
+            } else {
+                this.gatewayProxy.enableThroughputControl(throughputControlStore);
+            }
         }
 
         this.throughputControlStore.enableThroughputControlGroup(group, throughputQueryMono);