+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -10,6 +10,13 @@
 
 ### Other Changes
 
+## 1.10.1 (2023-09-10)
+
+### Other Changes
+#### Dependency Updates
+- Upgraded `azure-core` from `1.42.0` to version `1.43.0`.
+- Upgraded `azure-core-http-netty` from `1.13.6` to version `1.13.7`.
+
 ## 1.10.0 (2023-08-09)
 
 ### Features Added
@@ -18,6 +25,11 @@
 ### Other Changes
 - Renamed `enableSupportLogging` to `enableUnsafeSupportLogging`. This is a breaking change from 1.10.0-beta.1.
 - `DefaultAzureCredential` will try all developer credentials. Previously if a developer credential attempted to acquire a token and failed, it would stop the chain. Deployed credentials are unaffected.
+
+#### Dependency Updates
+- Upgraded `azure-core` from `1.41.0` to version `1.42.0`.
+- Upgraded `azure-core-http-netty` from `1.13.5` to version `1.13.6`.
+
 
 ## 1.10.0-beta.1 (2023-07-24)
 
@@ -87,9 +99,9 @@
 
 - Fixed a bug in managed identity not properly URLEncoding a value. [#34375](https://github.com/Azure/azure-sdk-for-java/pull/34375)
 
-### Other Changes	### Other Changes
+### Other Changes
 
-#### Dependency Updates	#### Dependency Updates
+#### Dependency Updates
 - Upgraded `azure-core` from `1.37.0` to version `1.38.0`.
 - Upgraded `azure-core-http-netty` from `1.13.1` to version `1.13.2`.
 - Upgraded `msal4j` from `1.13.5` to version `1.13.7`.