+++ sdk/resourcemanager/azure-resourcemanager-redis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 2.31.0-beta.1 (Unreleased)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Updated `api-version` to `2023-08-01`.
+
 ## 2.30.0 (2023-08-25)
 
 ### Other Changes