+++ sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 2.31.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 2.30.0 (2023-08-25)
 
 ### Bugs Fixed